techDebt(bulkUploadWorkersRefactor): Move validatePartTimeSalary tests into separate spec file
